+++ packages/contracts-ops/CHANGELOG.md
@@ -9,3 +9,4 @@
 - feature: callbatch contract
 - feature: Config.sol exposes configuration to foundry scripts
 - feature: Produce callbatches required to perform an Upgrade
+- feature: Reboot script & Reboot fork test setup